+++ src/components/RowHeader/RowHeader.tsx
@@ -1,4 +1,5 @@
-import { CSSProperties, MouseEvent, ReactNode } from 'react'
+import { CSSProperties, MouseEvent, ReactNode, useCallback, useRef } from 'react'
+import { useCellNavigation } from '../../hooks/useCellsNavigation'
 
 interface Props {
   busy?: boolean
@@ -7,19 +8,29 @@ interface Props {
   onClick?: (event: MouseEvent) => void
   showCheckBox?: boolean
   style?: CSSProperties
-  ariaColIndex?: number
+  ariaColIndex: number
+  ariaRowIndex: number
 }
 
-export default function RowHeader({ children, checked, onClick, showCheckBox, style, busy, ariaColIndex }: Props) {
+export default function RowHeader({ children, checked, onClick, showCheckBox, style, busy, ariaColIndex, ariaRowIndex }: Props) {
+  const ref = useRef<HTMLTableCellElement>(null)
+  const { tabIndex, navigateToCell } = useCellNavigation({ ref, ariaColIndex, ariaRowIndex })
+  const handleClick = useCallback((event: MouseEvent) => {
+    navigateToCell()
+    onClick?.(event)
+  }, [onClick, navigateToCell])
+
   const disabled = !onClick
   return (
     <th
+      ref={ref}
       scope="row"
       role="rowheader"
       style={style}
-      onClick={onClick}
+      onClick={handleClick}
       aria-busy={busy}
       aria-colindex={ariaColIndex}
+      tabIndex={tabIndex}
     >
       <span>{children}</span>
       { showCheckBox && <input type='checkbox' disabled={disabled} checked={checked} readOnly /> }