+++ src/components/TableCorner/TableCorner.tsx
@@ -1,4 +1,5 @@
-import { CSSProperties, MouseEvent, ReactNode } from 'react'
+import { CSSProperties, MouseEvent, ReactNode, useCallback, useRef } from 'react'
+import { useCellNavigation } from '../../hooks/useCellsNavigation'
 
 interface Props {
   checked?: boolean
@@ -6,12 +7,27 @@ interface Props {
   onClick?: (event: MouseEvent) => void
   showCheckBox?: boolean
   style?: CSSProperties
-  ariaColIndex?: number
+  ariaColIndex: number
+  ariaRowIndex: number
 }
 
-export default function TableCorner({ children, checked, onClick, showCheckBox, style, ariaColIndex }: Props) {
+export default function TableCorner({ children, checked, onClick, showCheckBox, style, ariaColIndex, ariaRowIndex }: Props) {
+  const ref = useRef<HTMLTableCellElement>(null)
+  const { tabIndex, navigateToCell } = useCellNavigation({ ref, ariaColIndex, ariaRowIndex })
+  const handleClick = useCallback((event: MouseEvent) => {
+    navigateToCell()
+    onClick?.(event)
+  }, [onClick, navigateToCell])
+
   return (
-    <td aria-disabled={!showCheckBox} style={style} onClick={onClick} aria-colindex={ariaColIndex}>
+    <td
+      ref={ref}
+      aria-disabled={!showCheckBox}
+      style={style}
+      onClick={handleClick}
+      aria-colindex={ariaColIndex}
+      tabIndex={tabIndex}
+    >
       <span>{children}</span>
       { showCheckBox && <input type='checkbox' checked={checked} readOnly /> }
     </td>