+++ test/helpers/dataframe.test.ts
@@ -140,46 +140,46 @@ describe('sortableDataFrame', () => {
   it('should return data sorted by column "age" in ascending order', async () => {
     const rows = await awaitRows(sortableDf.rows({ start: 0, end: 4, orderBy: [{ column: 'age', direction: 'ascending' as const }] }))
     expect(rows).toEqual([
-      { index: 2, cells:{ id: 2, name: 'Bob', age: 20 } },
-      { index: 3, cells:{ id: 4, name: 'Dani', age: 20 } },
-      { index: 0, cells:{ id: 3, name: 'Charlie', age: 25 } },
-      { index: 1, cells:{ id: 1, name: 'Alice', age: 30 } },
+      { index: 2, cells: { id: 2, name: 'Bob', age: 20 } },
+      { index: 3, cells: { id: 4, name: 'Dani', age: 20 } },
+      { index: 0, cells: { id: 3, name: 'Charlie', age: 25 } },
+      { index: 1, cells: { id: 1, name: 'Alice', age: 30 } },
     ])
   })
 
   it('should return data sorted by column "age" in descending order, using the data index in case of ties', async () => {
     const rows = await awaitRows(sortableDf.rows({ start: 0, end: 4, orderBy: [{ column: 'age', direction: 'descending' as const }] }))
     expect(rows).toEqual([
-      { index: 1, cells:{ id: 1, name: 'Alice', age: 30 } },
-      { index: 0, cells:{ id: 3, name: 'Charlie', age: 25 } },
-      { index: 2, cells:{ id: 2, name: 'Bob', age: 20 } },
-      { index: 3, cells:{ id: 4, name: 'Dani', age: 20 } },
+      { index: 1, cells: { id: 1, name: 'Alice', age: 30 } },
+      { index: 0, cells: { id: 3, name: 'Charlie', age: 25 } },
+      { index: 2, cells: { id: 2, name: 'Bob', age: 20 } },
+      { index: 3, cells: { id: 4, name: 'Dani', age: 20 } },
     ])
   })
 
   it('should return data sorted by columns "age" in ascending order and "name" in descending order', async () => {
     const rows = await awaitRows(sortableDf.rows({ start: 0, end: 4, orderBy: [{ column: 'age', direction: 'ascending' as const }, { column: 'name', direction: 'descending' as const }] }))
     expect(rows).toEqual([
-      { index: 3, cells:{ id: 4, name: 'Dani', age: 20 } },
-      { index: 2, cells:{ id: 2, name: 'Bob', age: 20 } },
-      { index: 0, cells:{ id: 3, name: 'Charlie', age: 25 } },
-      { index: 1, cells:{ id: 1, name: 'Alice', age: 30 } },
+      { index: 3, cells: { id: 4, name: 'Dani', age: 20 } },
+      { index: 2, cells: { id: 2, name: 'Bob', age: 20 } },
+      { index: 0, cells: { id: 3, name: 'Charlie', age: 25 } },
+      { index: 1, cells: { id: 1, name: 'Alice', age: 30 } },
     ])
   })
 
   it('should slice the sorted data correctly in ascending order', async () => {
     const rows = await awaitRows(sortableDf.rows({ start: 1, end: 3, orderBy: [{ column: 'id', direction: 'ascending' as const }] }))
     expect(rows).toEqual([
-      { index: 2, cells:{ id: 2, name: 'Bob', age: 20 } },
-      { index: 0, cells:{ id: 3, name: 'Charlie', age: 25 } },
+      { index: 2, cells: { id: 2, name: 'Bob', age: 20 } },
+      { index: 0, cells: { id: 3, name: 'Charlie', age: 25 } },
     ])
   })
 
   it('should slice the sorted data correctly in descending order', async () => {
     const rows = await awaitRows(sortableDf.rows({ start: 1, end: 3, orderBy: [{ column: 'id', direction: 'descending' as const }] }))
     expect(rows).toEqual([
-      { index: 0, cells:{ id: 3, name: 'Charlie', age: 25 } },
-      { index: 2, cells:{ id: 2, name: 'Bob', age: 20 } },
+      { index: 0, cells: { id: 3, name: 'Charlie', age: 25 } },
+      { index: 2, cells: { id: 2, name: 'Bob', age: 20 } },
     ])
   })